move structs to schema directory
