+++ backend-go/database/store_acts.go
@@ -3,7 +3,7 @@ package database
 import (
 	"context"
 
-	"github.com/areeeeeeeb/reLive/backend-go/models"
+	"github.com/areeeeeeeb/reLive/backend-go/schema/models"
 	"github.com/jackc/pgx/v5"
 )
