+++ backend-go/database/store_users.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"errors"
 
-	"github.com/areeeeeeeb/reLive/backend-go/models"
+	"github.com/areeeeeeeb/reLive/backend-go/schema/models"
 	"github.com/jackc/pgx/v5"
 )
